feat: 요청 핸들러 삭제 후 홈, 로그인, 회원가입, 리소스 컨트롤러 구현
