+++ tomcat/src/main/java/nextstep/jwp/presentation/HomePageController.java
@@ -1,5 +1,6 @@
-package org.apache.coyote.handler;
+package nextstep.jwp.presentation;
 
+import org.apache.catalina.controller.AbstractController;
 import org.apache.coyote.request.HttpRequest;
 import org.apache.coyote.response.HttpResponse;
 import org.apache.coyote.response.ResponseBody;
@@ -8,20 +9,16 @@ import static org.apache.coyote.common.HttpVersion.HTTP_1_1;
 import static org.apache.coyote.common.MediaType.TEXT_HTML;
 import static org.apache.coyote.response.HttpStatus.OK;
 
-public class HomeRequestHandler implements RequestHandler {
-
-    private static final String DEFAULT = "Hello world!";
+public class HomePageController extends AbstractController {
 
     @Override
-    public HttpResponse handle(final HttpRequest httpRequest) {
-        final ResponseBody responseBody = new ResponseBody(DEFAULT);
+    protected void doGet(final HttpRequest request, final HttpResponse response) throws Exception {
+        final ResponseBody responseBody = ResponseBody.from("Hello world!");
 
-        return HttpResponse.builder()
-                .setHttpVersion(HTTP_1_1)
+        response.setHttpVersion(HTTP_1_1)
                 .setHttpStatus(OK)
                 .setContentType(TEXT_HTML.value())
                 .setContentLength(responseBody.length())
-                .setResponseBody(responseBody)
-                .build();
+                .setResponseBody(responseBody);
     }
 }
