+++ tomcat/src/main/java/nextstep/jwp/presentation/ResourceController.java
@@ -1,5 +1,6 @@
-package org.apache.coyote.handler;
+package nextstep.jwp.presentation;
 
+import org.apache.catalina.controller.AbstractController;
 import org.apache.coyote.common.MediaType;
 import org.apache.coyote.request.HttpRequest;
 import org.apache.coyote.response.HttpResponse;
@@ -9,20 +10,18 @@ import org.apache.coyote.util.ResourceReader;
 import static org.apache.coyote.common.HttpVersion.HTTP_1_1;
 import static org.apache.coyote.response.HttpStatus.OK;
 
-public class ResourceRequestHandler implements RequestHandler {
+public class ResourceController extends AbstractController {
 
     @Override
-    public HttpResponse handle(final HttpRequest httpRequest) {
-        final String requestPath = httpRequest.requestLine().requestPath().value();
+    protected void doGet(final HttpRequest request, final HttpResponse response) throws Exception {
+        final String requestPath = request.requestLine().requestPath().value();
         final String resourceBody = ResourceReader.read(requestPath);
-        final ResponseBody responseBody = new ResponseBody(resourceBody);
+        final ResponseBody responseBody = ResponseBody.from(resourceBody);
 
-        return HttpResponse.builder()
-                .setHttpVersion(HTTP_1_1)
+        response.setHttpVersion(HTTP_1_1)
                 .setContentType(MediaType.from(requestPath).value())
                 .setHttpStatus(OK)
                 .setResponseBody(responseBody)
-                .setContentLength(resourceBody.length())
-                .build();
+                .setContentLength(resourceBody.length());
     }
 }
