+++ tomcat/src/test/java/nextstep/jwp/presentation/HomePageControllerTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.handler.get;
+package nextstep.jwp.presentation;
 
 import org.apache.coyote.HttpFormTestUtils;
 import org.apache.coyote.Processor;
@@ -8,21 +8,21 @@ import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import support.StubSocket;
 
-import java.io.IOException;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class UserRegisterRequestGetHandlerTest {
+class HomePageControllerTest {
 
     @Test
-    void 파일_확장자를_알_수_없으면서_GET일_경우_회원가입_페이지를_응답한다() throws IOException {
+    void HelloWorld_페이지를_응답한다() {
         // given
         final String httpRequest = HttpFormTestUtils.builder()
-                .GET().requestUri("/register").http11().enter()
+                .GET().requestUri("/").http11().enter()
                 .host("localhost:8080").enter()
                 .connection("keep-alive").enter()
+                .contentLength("0").enter()
+                .accept("text/html;charset=utf-8").enter()
                 .enter()
                 .build();
 
@@ -35,9 +35,9 @@ class UserRegisterRequestGetHandlerTest {
         // then
         final String httpResponse = HttpFormTestUtils.builder()
                 .http11().OK().enter()
-                .contentLength("4319").enter()
+                .contentLength("12").enter()
                 .contentType("text/html;charset=utf-8").enter()
-                .enter().responseByResource("static/register.html")
+                .enter().responseBody("Hello world!")
                 .build();
 
         assertThat(socket.output()).isEqualTo(httpResponse);
