+++ tomcat/src/test/java/nextstep/jwp/presentation/RegisterPageControllerTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.handler.post;
+package nextstep.jwp.presentation;
 
 import org.apache.coyote.HttpFormTestUtils;
 import org.apache.coyote.Processor;
@@ -8,11 +8,40 @@ import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import support.StubSocket;
 
+import java.io.IOException;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class UserRegisterRequestPostHandlerTest {
+class RegisterPageControllerTest {
+
+    @Test
+    void 파일_확장자를_알_수_없으면서_GET일_경우_회원가입_페이지를_응답한다() throws IOException {
+        // given
+        final String httpRequest = HttpFormTestUtils.builder()
+                .GET().requestUri("/register").http11().enter()
+                .host("localhost:8080").enter()
+                .connection("keep-alive").enter()
+                .enter()
+                .build();
+
+        final StubSocket socket = new StubSocket(httpRequest);
+        final Processor processor = new Http11Processor(socket);
+
+        // when
+        processor.process(socket);
+
+        // then
+        final String httpResponse = HttpFormTestUtils.builder()
+                .http11().OK().enter()
+                .contentLength("4319").enter()
+                .contentType("text/html;charset=utf-8").enter()
+                .enter().responseByResource("static/register.html")
+                .build();
+
+        assertThat(socket.output()).isEqualTo(httpResponse);
+    }
 
     @Test
     void 요청_바디에_회원_정보를_넣어_회원가입_요청을_성공하면_메인_홈페이지로_리다이렉트한다() {
